# src/main/resources/requests/Historia-Medica_Swagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/usuarios/profesionales/{id}:
    put:
      tags:
      - usuario-controller-rest
      operationId: updateProfesional
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioUpdateRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
...
  /api/profesionales/{id}/removerEspecialidades:
    put:
      tags:
      - profesional-controller-rest
      operationId: removeEspecialidadesInProfesional
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfesionalEspecialidadesRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/profesionales/{id}/removerEspecialidad/{especialidadId}:
    put:
      tags:
      - profesional-controller-rest
      operationId: removeEspecialidadInProfesional
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: especialidadId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProfesionalResponse'
  /api/profesionales/{id}/agregarEspecialidades:
    put:
      tags:
      - profesional-controller-rest
      operationId: addEspecialidadesInProfesional
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfesionalEspecialidadesRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/profesionales/{id}/agregarEspecialidad/{especialidadId}:
    put:
      tags:
      - profesional-controller-rest
      operationId: addEspecialidadInProfesional
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: especialidadId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProfesionalResponse'
...
  /api/usuarios/{id}/nuevoContacto:
    post:
      tags:
      - usuario-controller-rest
      operationId: addContacto
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactoRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
  /api/usuarios/{id}/nuevaDireccion:
    post:
      tags:
      - usuario-controller-rest
      operationId: addDireccion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DireccionRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
...
  /api/register/profesionales:
    post:
      tags:
      - usuario-controller-rest
      operationId: createProfesional
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioProfesionalRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/profesionales:
    get:
      tags:
      - profesional-controller-rest
      operationId: getList_2
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageProfesionalResponse'
    post:
      tags:
      - profesional-controller-rest
      operationId: createProfesional_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfesionalRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProfesionalResponse'
  /api/profesionales/{id}/nuevoContacto:
    post:
      tags:
      - profesional-controller-rest
      operationId: addContacto_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactoRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProfesionalResponse'
  /api/profesionales/{id}/nuevaDireccion:
    post:
      tags:
      - profesional-controller-rest
      operationId: addDireccion_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DireccionRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProfesionalResponse'
  /api/prescripciones:
    get:
      tags:
      - prescripcion-controller-rest
      operationId: getList_3
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagePrescripcionResponse'
    post:
      tags:
      - prescripcion-controller-rest
      operationId: createPrescricpion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrescripcionRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PrescripcionResponse'
...
  /api/especialidades:
    post:
      tags:
      - profesional-controller-rest
      operationId: createEspecialidad
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EspecialidadRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EspecialidadResponse'
...
  /api/centrosSalud/{id}/nuevoContacto:
    post:
      tags:
      - centro-salud-controller-rest
      operationId: addContacto_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactoRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CentroSaludResponse'
  /api/centrosSalud/{id}/nuevaDireccion:
    post:
      tags:
      - centro-salud-controller-rest
      operationId: addDireccion_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DireccionRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CentroSaludResponse'
...
  /api/usuarios:
    get:
      tags:
      - usuario-controller-rest
      operationId: getList
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageUsuarioResponse'
...
  /api/usuarios/profesionales:
    get:
      tags:
      - usuario-controller-rest
      operationId: getListProfesionales
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageUsuarioProfesionalResponse'
...
  /api/profesionales/{id}/especialidades:
    get:
      tags:
      - profesional-controller-rest
      operationId: getEspecialidadesByProfesional
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/historiasMedicas/{id}/profesionales:
    get:
      tags:
      - profesional-controller-rest
      operationId: getProfesionalesByHistoriaMedica
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
...
    ProfesionalResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        direcciones:
          type: array
          items:
            $ref: '#/components/schemas/DireccionResponse'
        contactos:
          type: array
          items:
            $ref: '#/components/schemas/ContactoResponse'
        especialidades:
          type: array
          items:
            $ref: '#/components/schemas/EspecialidadResponse'
...
    UsuarioUpdateRequest:
      type: object
      properties:
        nombre:
          type: string
        apellido:
          type: string
        mail:
          type: string
        fechaNacimiento:
          type: string
...
    UsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        apellido:
          type: string
        fechaNacimiento:
          type: string
        mail:
          type: string
        genero:
          type: string
        estado:
          type: boolean
        direcciones:
          type: array
          items:
            $ref: '#/components/schemas/DireccionResponse'
        contactos:
          type: array
          items:
            $ref: '#/components/schemas/ContactoResponse'
        autorizacion:
          $ref: '#/components/schemas/AutorizacionComponentResponse'
...
    ProfesionalEspecialidadesRequest:
      type: object
      properties:
        especialidadesIds:
          type: array
          items:
            type: integer
            format: int64
...
    ContactoRequest:
      type: object
      properties:
        telefono:
          type: string
        mailAlternativo:
          type: string
    DireccionRequest:
      type: object
      properties:
        direccion:
          type: string
        piso:
          type: integer
          format: int32
        departamento:
          type: string
        referencia:
          type: string
        idProvincia:
          type: integer
          format: int64
        idLocalidad:
          type: integer
          format: int64
...
    UsuarioProfesionalRequest:
      type: object
      properties:
        nombre:
          type: string
        apellido:
          type: string
        mail:
          type: string
        fechaNacimiento:
          type: string
        genero:
          type: string
        password:
          type: string
        matricula:
          type: string
...
    ProfesionalRequest:
      type: object
      properties:
        nombre:
          type: string
        idHistoriaMedica:
          type: integer
          format: int64
...
    EspecialidadRequest:
      type: object
      properties:
        nombre:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PageProfesionalResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProfesionalResponse'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        pageable:
          $ref: '#/components/schemas/PageableObject'
        last:
          type: boolean
        empty:
          type: boolean
    PagePrescripcionResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/PrescripcionResponse'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        pageable:
          $ref: '#/components/schemas/PageableObject'
        last:
          type: boolean
        empty:
          type: boolean
...